styling refactor to includes styles module
